requirements.txt
aiohttp>=3.9
python-dotenv>=1.0
tzdata>=2024.1
fastapi>=0.135
uvicorn[standard]>=0.44
websockets>=16.0
aiosqlite>=0.22
pydantic>=2.9
